requirements.txt
matplotlib==3.9.2
mplcyberpunk==0.7.5
pandas==2.2.3
numpy==2.1.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10